requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.7
django-summernote==0.8.20.0
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
webencodings==0.5.1
